chore(code-garden): split ContentSchedulerTab into hook + sub-components [2026-W30]
Non-functional change from repo audit 2026-W30.
Finding: [Medium] ContentSchedulerTab.jsx monolith grew from 450 → 637 lines.

Extracts the data/handler layer into useContentScheduler.js, the item card
into SchedulerItemCard.jsx, and constants/date helpers into
contentSchedulerConstants.js. The main ContentSchedulerTab.jsx is now a slim
container (composer + calendar grid + unscheduled strip + day-strip).

All 175 mission-control tests pass (15 ContentSchedulerTab tests).
AC met: no sub-file exceeds 250 lines.
