+++ docs/repo-audits/2026-W30.md
@@ -341,7 +341,7 @@ No new performance concerns this week. The Content Scheduler auto-post correctly
 
 ### Milestone 2 — High-leverage improvements
 
-**2-A. Extract `useContentScheduler` hook + row/calendar sub-components from `ContentSchedulerTab.jsx`**
+**2-A. Extract `useContentScheduler` hook + row/calendar sub-components from `ContentSchedulerTab.jsx`** ✅ [PR #TBD](https://github.com/Stoffer-Industries/sindustries/pull/TBD)
 - *Description:* Same decomposition target as W29 Theme 4, now larger (637 lines).
 - *AC:* No sub-file exceeds 250 lines.
 - *Effort:* M
